写真選択時の制御など追加

+++ VRChatToolBox/LogEditor.cs
@@ -81,14 +81,20 @@ namespace VRChatToolBox
         // 投稿後のメタデータの移動
         internal static void MoveMetaDataFile(string filePath)
         {
+            // 万が一メタデータが無ければ返す
+            if (!File.Exists(filePath)) return;
+
             string fileName = Path.GetFileName(filePath);
             string destPath = $"{ProgramSettings.Settings.DesignatedUpLoadedInfoPath}\\{fileName}";
+
             // 既にあれば移動しない
             if (File.Exists(destPath)) return;
             // なければフォルダを作る
             if (!Directory.Exists(ProgramSettings.Settings.DesignatedUpLoadedInfoPath))
                  Directory.CreateDirectory(ProgramSettings.Settings.DesignatedUpLoadedInfoPath);
+
             File.Move(filePath, destPath);
+
         }
     }
 }
